Comment out AdminstrationNode in config file
In txnvalidator.js, AdminstrationNode was improperly set
to a node identifier based on a testing configuration. Since
node identifiers are not known until a network is established,
we should not provide a default value here. It is also not
meaningful to set this to the node's own identifier, as nodes
do not send themselves shutdown messages.

This commit comments out the configuration of AdminstrationNode
to present it as an example. The behavior in txnvalidator will
be that no administration node is set, and thus the node will not
process shutdown messages. This is equivalent to the current state
of having an invalid AdminstrationNode identifier.