Fix Sphinx warnings by qualifying class attributes in docstrings
This is to work around a defect in Sphinx with the way it processes the :ivar directive.  The :attribute does not appear to exhibit this defect. However, we decided to have napoleon generate :ivar rather than :attribute (see "napoleon_use_ivar = True" in sawtooth-docs/source/conf.py) because the rendered document looks better.