Separate console verbosity from logging
This change removes support for __screen__ as a logging destination in
favor of a dedicated verbosity switch for logging to the console.

The use of '-v' instead of specifying the specific logging level is
intended to allow us to make decisions about what should be logged as
the verbosity increases (via multipe '-v' switches).  Currently, this
just increases from WARN (default), INFO (-v), DEBUG (-vv).

In parallel with console logging, file logging continues; since we can
control the verbosity of the console separately, a default of DEBUG for
file logging is probably always reasonable.