Break logging dependency between mktplace and validator
Previously, mktplace used txnserver.log_setup.setup_loggers to
initialize the mktplace CLI.  Now, that code is duplicated into this
repo so that mktplace has its own copy.  This code is expected to
deviate from txnserver since logging for a CLI and logging for a server
process have different characteristics.