Add ability to capture stdout/stderr to logger setup
When we run in daemon mode, stdout and stderr are closed.  A normal
approach is to redirect these to /dev/null.  However, fatal errors
(via warn.warning) are output to stderr.  Therefore, it is desirable
to capture and redirect them to the log (via the logger).

This implementation of capturing stdout and stderr is remedial at
best.  It will only work for high-level python code (the Enclave
module stdout/stderr will not be captured).  It also does not break
on newlines (though it will naively skip empty lines), and because the
division is on the calls to write(), it is very possible that output
will be split across multiple log entries in odd ways.

A good implementation requires threading (for example, to read the
other side of a pipe to send it to the logger).  Given the low
importance of this, the simple solution provided here is probably
good enough with the least potential for side effects.