Remove dependency between txnserver.config and gossip
This removes the use of json2dict for loading JSON-like config
files.

While this should be mostly a non-functional change, this does remove
the explicit conversion to ascii done in json2dict for config file
loads.  The assumption is that the conversion is not actually necessary
(or desirable) here.

The motivation for breaking the dependency between sawtooth.config
and gossip is to prevent pulling in the twisted framework when it
is otherwise unnecessary.  (Imports of twisted have epoll
side-effects which conflict with use of fork().)