Remove Validator.LedgerWebClient attribute
This attribute no longer makes sense now that multiple URLs are
supported.  As it was, it was set to the client associated with
the last url in the loop, but never used in code outside the
loop.  This passes the url into get_endpoint() and creates the
client there instead.