Replace config save/load/init with keygen and args
This removes the concept of a config file from the tool.  Instead of
a 'sawtooth init', we now have 'sawtooth keygen' which is more
focused and will replace txnkeygen.

Submit also takes a --key argument instead of --username, since not
all transaction families may have the concept of a username as such.
The key can be either a PATH to a key file or just the name of the
key; if it is just the name, it is assumed to be at
~/.sawtooth/KEY.wif.