Switch test shutdown from post message to send SIGINT
In order to support dumping profiler information we need
a handled shutdown of the validator code. For some reason,
the previous approach of posting a network shutdown message
to the validators puts the system into a state where the
SIGTERM/SIGINT handlers for twisted reactor do not function
correctly. This change eliminates the network shutdown message
in favor of a two-stage SIGINT/SIGKILL process. While there
are still issues in the shutdown handling which require the
SIGKILL, this change does cause the reactor SIGINT handler to
run.