new 'quorum advertisement'
Node will continually rebroadcast itself until it is in the endpoint
registry.  The purpose of being in the endpoint registry in quorum is
simply transparency, as participation in the quorum process is all
derived from configs.  The problem with the old mechanism was that there
had to be a quorum to get into the endpoint registry, which obviously
complicates bootstrapping in a known topology.

It would be best if this mechanism were moved out of this high traffic
area.  We suggest modifying quorum advertisement to handle this issue,
presuming we want to continue to use the endpoint registry (in quorum).
The author has not do so because the code in this branch was rigorously
tested, and that level of change would require a significant amount of
retesting.