add parameters to test infrastructure and test smoke
Timeouts should be variable to accommodate different quorum mechanisms.

Its generally nice to parameterize the number of rounds as well.  In
particular, we need to because quorum currently lacks a catch up
protocol and possibly has some synchronization issues.  We currently
know we can pass the initial tests, and are interested in maintaining
that condition until we can circle back to address outstanding issues
with quorum.