Reduce serialization when using 'ThreadSafeShelf'
Modifies 'ThreadSafeShelf' to be derived from object (thus
ensuring we haven't missed a method in our locking).

Caches objects in a dict and returns those with preference to avoid
serialization.  Attempts to keep backward compatiblity with the shelf
behavior by also maintaining a shelf object with the same data.

This substantially improves performance of journal_core as it reduces
the amount of signature verifications, which were recurring every time
an object was deserialized.