Add --family, -F flag to txnvalidator
Added flags --family and -F to txnvalidator, and tested.
This is part of NML-1122. This commit does not change the
tutorial, which will be a separate commit.
Although the code change was tested, reviewer should make
sure that no side effects are caused by this change, as I
am new to the code base.