Update n2i in mktplace globalobjectstore to take an object type
The object model now includes 'object-id' and 'full-name' attributes.
The full-name is creatorID/name e.g. 483084ndif0ih4/account/USD if it isn't
a participant. If the object is a participant it is //username. Because of
how lookup works an object-type needs to be passed into n2i. This results
in a few places in the cli to include --type.

Signed-off-by: Boyd Johnson <boydx.johnson@intel.com>