genesis tool initial draft
sawtooth genesis provides us with a framework to obsolete --genesis and
--restore flags by preparing the genesis block for the primordial
validator offline.

USAGE:
    0)  create a config file <CFG> with the following properties:
            {
                'GenesisLedger': False,
                'Restore': True,
                'LedgerURL': [],
                'InitialConnectivity': 0,
                'KeyFile': <KEYFILE>,
            }
    1)  make sure the key file exists
        --  ./bin/sawtooth keygen <KEYFILE>
    2)  create the block:
        --  ./bin/sawtooth genesis -vv --config <CFG>
    3)  start the ledger:
        --  ./bin/txnvalidator -vv --config <CFG>

You should be able to observe that the restore-oriented commit head from
(3) matches the head reported in (2)

During normal operation, InitialConnectivity must be 0 in order to allow
our validator to reload its own ledger (including our one-off,
offline-generated genesis block), as ledgers and validators are
currently tightly coupled.  If this is not set to 0, the (now
non-genesis) validator will loop in an attempt to discover peers
(preventing proper bootstrapping).  Similarly, LedgerURL must be an
empty list.

Signed-off-by: Gregory Setterholm <gregoryx.setterholm@intel.com>