Integrate MarketPlace Navigator
MarketPlace Navigator is a hosted UI for the MarketPlace Transaction
Family.  It consists of three major pieces: a single page app, a server
backend, and a syncing tool for synchronizing block state with the
navigator's database.

The Single Page App

The single page app is a ClojureScript app (with some plain JavaScript
dependencies), which provides client-side transaction signing capabilities.
A user of the system generates a WIF key in the browser, which they can save
localy. A user can create most MarketPlace objects via the UI, including
participants, assets, holdings and sell offers.

The Server Backend

The server is a Node.js application, which uses RethinkDB for its
storage needs.  It provides API's to create richer representation of the
marketplace objects, in order to improve the user experience. It also
acts as a pass-through for transactions being submitted to a validator.
Any transaction successfully submitted will be recorded, giving history
of any failures that may occur during block construction.

The Synchronization Tool

The synchronization tool is a python script that bridges the gap between
block store contents and a RethinkDB database instance.  It reads block
deltas from a validator, and creates block store snapshots, essentailly
providing an immutable history of the block state.

Setup

Use the plugins in <sawtooth-core>/tools/plugins for `install-rethink` and
`install_js_tools`.  Instructions for building and running each of the
major pieces is in the READMEs provided.

Signed-off-by: Peter Schwarz <peterx.schwarz@intel.com>