Change block list command to return all blocks by default
The block list command had the option --all, which would return all
blocks, and was not the default. By default, the command returned
the default of the --blockcount option, which was 10.

Changed the command to remove the --all option, and cause the
command to return all blocks by default, unless the --blockcount
option was set to a value. The --blockcount parameter is used to
pass the integer -1 by default now, to signal that all blocks
should be returned.