remove validator subclasses
By passing a ledger object to a validator, we can get rid of most
validator subclasses, since they had already been reduced to:
    --  the class variable EndpointDomain (which we have made instance)
    --  a trivial init method
    --  an initialize_ledger_from_node method

A ledger object currently needs node information, so we have broken node
(and http_port) determination into a standalone function.  This way, we
can create the node, from which we can create the ledger (also in a
standalone manner).  This detatched ledger can then be passed to the
validator constructor.

The QuorumValidator subclass remains.  Cursory analysis shows that it
is doing some different things with respect to startup and topological
updates.  The difference with respect to startup is arguably topologic
in nature, suggesting that a topology module/interface might be
advantageous.

Signed-off-by: Gregory Setterholm <gregoryx.setterholm@intel.com>