Refactor Register, Unregister baseclasses into util functions
Register and Unregister have three checks in is_valid,
a check on whether the name is valid, a check on whether the
creator of the object sent the transaction, and whether the
ObjectId is already in the store. These are refactored into
util functions and for the case of the last test pushed into the
former subclass.

Signed-off-by: BoydX Johnson <boydx.johnson@intel.com>