block integration test stub
We'll use this test to start building tall.

The immediate goal is to get the existing low level constructs
integrated with the higher levels, so that we can know the blocks we
create (e.g. the genesis block) can be consumed by a running validator.

We can then use this test to demonstrate to ourselves that as we add
poet1 features and/or specification requirements, they work at the macro
level.

Signed-off-by: Gregory Setterholm <gregoryx.setterholm@intel.com>