non-functionals: manage resources outside launcher
Now that we have a genesis block utility, we want to be able to control
the allocation and deallocation of resources needed by both the genesis
util and its corresponding initial validator outside of the launcher.

This patch further reduces the launcher's cognizance of and
responsibility for the configuration of the validators it launches, as
well as shifts the power for creating and deallocating the resources
that might be needed to satisfy a configuration from the launcher
(ValidatorCollectionController) over to:
    a)  the NetworkConfig class (for configuration management)
    b)  a provider class, intended to manage resources for NetworkConfig
        in a pluggable fashion

Now that configs can be generated and modified before launching a node,
the genesis util can take advantage of the same test framework as the
launcher.

Additionally, the startup tests should be trivial to repair with this
new framework.

Signed-off-by: Gregory Setterholm <gregoryx.setterholm@intel.com>