copy missing elements from VNM to VCC
We've recently improved out shutdown capabilities, and this ports those
improvements over to our new infrastructure.  This commit just copies these
methods from VNM to VCC, so that alterations can be examined in the next
commit.

We're additionally grabbing unpack blockchain for future uses as that
seems to be an important argument for launcher.

Signed-off-by: Gregory Setterholm <gregoryx.setterholm@intel.com>