Fix file used by tutorial after --genesis flag removal
Minor change to a new bash script that is run as part of the
tutorial due to the --genesis and --restore flag removal work.

The bash script invokes the new poet-genesis utility in sawtooth CLI.
This command needs to specify the integer key transaction family.
It was originally left out, because I thought that this was part
of the default config. Testing showed this not to be true.