Add prototype python sawtooth_sdk.client.stream package
This package handles the client-side of the grpc-based messaging
system.  This is a functional, but rough, prototype.  In particular,
it does not yet cleanly shutdown elegantly.  The python grpc library
is not necessarily a friendly API for the purposes of implementing
bi-directional messaging, and as such, the code is more complex
than seems necessary.

Signed-off-by: Shawn T. Amundson <amundson@bitwise.io>