NodeConfig class
In order to be able to launch nodes in wrapped mode, we need a less
restrictive method of starting nodes than that based solely on node
name, port and http port.  However, we still want to preserve the
simplicity of using these three arguments in many cases.

Our solution is to have node controllers and node command generators
operate on a new NodeConfig base class, when starting nodes.  NodeConfig
objects can be simple where they need to be, but have increased
flexibility in subclass implementations.

Signed-off-by: Gregory Setterholm <gregoryx.setterholm@intel.com>