Implement java tp client code
The MessageListener handles all incoming messages, either updating the Futures
if the message is a response to a sent message, or putting the message on the
receiveQueue. To get that message any client using Stream would call Stream().receive();
The State class handles interactions with the validator ContextManager using Stream.

Signed-off-by: BoydX Johnson <boydx.johnson@intel.com>