Update bin/run_lint for correctness and to take options
This changes the way run_lint calls pep8, pylint, and file
commands.  A list of packages is now determined in each of
the subdirectories we check and the packages are fed to the
commands.  The list of packages can be substantially
reduced by providing the -s option with the originating
branch (usually upstream/master); in this case, the command
will attempt to only lint files which have changed.

Other new options include -v (verbose).  One -v gives
high-level section output, and -vv turns on -x in bash,
given substantial debugging information.  Also, a -n (dryrun)
option is added which will show which files will be
linted. A -h (help) option is also provided.

Signed-off-by: Shawn T. Amundson <amundson@bitwise.io>