remove launcher_cli host/endpoint args
This commit assumes that the host argument by itself is not useful for the
launcher cli, and was only being used to differentiate from the endpoint host
in order to expose http on all interfaces.

If SubprocessNodeController crafts its listen directives to use 0.0.0.0 for
http and localhost for gossip, then the primary motivation for these arguments
(being able to reach the web interface from e.g. outside of vagrant) are
satisfied by default.

Signed-off-by: Gregory Setterholm <gregoryx.setterholm@intel.com>