Update interface of scheduler for keeping track of batches.
The TxnInformation has a txn (Transaction protobuf) and
new_state_hash and inform (booleans). The new_state_hash tells
the Executor that a new state hash needs to be made before the next
txn's context is made, while the inform says that the end of a batch
has happened and so all the txns in between should be checked and the
BatchStatus with valid (boolean) and state_hash (str) should be set.

Signed-off-by: Boyd Johnson <boydx.johnson@intel.com>