Workaround for time sync issue during vagrant up
Previously, we built a few RPMs in /project and then cached the
results.  At some point (maybe the switch to newer VirtualBox
or Ubuntu 16.04), the time within the guest is no longer in sync
with the host os enough to build debian packages such as
pybitcointools.

This change dues the build in /tmp but keeps the caching mechanism
as-is.  In the future, we may remove the caching so we always get
predictable results (despite the extra time).