Add sawtooth config transaction family.
The sawtooth config transaction family handles putting configuration
values into a validator's state.

Depending on the
`sawtooth.config.authorization_type` setting's value, additional settings
are either applied directly (in the case of 'None') or applyed after a
series of votes.

Settings transaction content is made up of protobuf messages, as well as
the storage format in the state is also a protobuf message.

Additionally:

* core_transactions/config will be linted

Signed-off-by: Peter Schwarz <peterx.schwarz@intel.com>