Add libsecp256k1 based signer
This commit adds a sawtooth signer implementation backed by bitcoin's
libsecp256k1 library via the secp256k1 python bindings.

To use this signer instead of the existing one simply change imports. This
commit does not change the imports of the consumers so as to enable some
trial by all the signer consumers.

> from sawtooth_signing import secp256k1_signer as signer

Also, as the project is between python2 and python3 at the moment, there
are code paths for both. It is intended that the python2 code will be
removed when the cutover is complete for the sawtooth 0.8 architecture.

Signed-off-by: dcmiddle <dan.middleton@intel.com>