Add GenesisController to apply GenesisData
The GenesisController is utilized on startup to determine if

a. The validator requires the creation of a genesis block
b. The validator has the correct set of files for either the genesis
process or regular execution

In the case of creating a genesis block, the GenesisController processes
the batches supplied in the serialized GenesisData and uses them to
create the genesis block.  On completion, the serialized genesis data is
deleted, the resulting block id is writen to
`<SAWTOOTH_DATA_DIR>/network-id` and the standard services are started
normally.

Any error in processing result in a shutdown of the validator.

Given that the validator will require the `sawtooth_config` transaction
familty, at a minimum, execution of transactions (at any time) is
paused until the transaction processor is connected. This requires that
validators to be launched in conjunction with tp_config.  Integration
tests reflect this change.