Implement asynchronous processing dispatcher
The dispatcher has handlers added to it for message types.
If it has multiple handlers for the same message type it executes
the handlers in the order that they were added, asynchronously.
Handlers implement the Handler abstract base class and return
a HandlerResult that must have a HandlerStatus.

Added unit test that tests that no matter what happens inside
a particular handler -- in this case time.sleep of a variable
amount of time -- the identity and message that were added to
the dispatcher are returned to the zmq interface.

Signed-off-by: Boyd Johnson <boydx.johnson@intel.com>