Streamline zmq interfaces into one class
_SendReceive uses zeromq sockets(either Dealer or Router sockets)
to send and receive msgs from internal components and from
other validator nodes. Interconnect takes an instance of _SendReceive
and executes it in a long running thread specifying that it
should use either a Dealer or a Router socket. Router sockets are
used used for the internal component interface. Each Interconnect
instance (specifically the external interface) that has peer
connections will start up Connection (zeromq Dealer sockets) and
make connections with other validator's Router sockets. This is
referred to as the zeromq harmony pattern.

Signed-off-by: Boyd Johnson <boydx.johnson@intel.com>