Add support for secured ZMQ sockets
The network instance of Interconnect has security turned on and
uses a hardcoded public/private key pair to identify the server
socket (this is the same on each validator). These keys are only
used during handshake, and are not the keys used to encrypt the
transmission. Client sockets generate ephemeral keys for the purpose
of authentication of unique connections to the server and so that
the exchange of transient keys to secure the conversation can take
place.

Inter-component interconnect has security turned off by default.

Signed-off-by: James Mitchell <mitchell@bitwise.io>