Implement logic for waiting or failing transactions
The TransactionExecutor will read from config state the list of
required transaction processors. If there are no required ones,
it will wait on each transaction that doesn't have a registered
transaction processor. This is useful on startup when configuration
hasn't happened yet. If there are required tps, any transaction that
comes in that isn't on the list will be failed immediately.

Signed-off-by: Boyd Johnson <boydx.johnson@intel.com>