Update to wait for test container to exit
Previously the run_docker_test script used --abort-on-container-exit
to stop all containers after the first one exited. It made the assumption
that the first container to exit would be the test container. This change
allows for stopping processes within containers and allowing the container
to exit, but allowing the test container (and all other containers) to
keep running.

Signed-off-by: Boyd Johnson <boydx.johnson@intel.com>