Modify error handling in validator main()
Modifies the try/except block to except all Exceptions and uses
LOGGER.error() instead of printing directly to stderr.  The logging
system should have been initialized prior to this try/except, and
it seems desirable that these error messages end up in the log file
in addition to console output (which is on by default now for WARN
and ERROR).

Signed-off-by: Shawn T. Amundson <amundson@bitwise.io>