Make Completer._process_incomplete_blocks() process all available work
Block dependencies may be long, especially in the case of initial load
for a new validator back to the genesis block. When a dependency is
satisfied, make sure that all outstanding work is processed in the
dependency chain. This is done iteratively because of the likelihood
of overflowing the default Python stack if done recursively.

Signed-off-by: James Mitchell <mitchell@bitwise.io>