Add Jenkinsfile whitelist
This commit adds a stage to the pipeline to check whether the user who
submitted the PR is on a whitelist. If they are not, the pipeline stops
immediately to prevent execution of untrusted code. This step is in
addition to the Verify Scripts stage, which prevents execution of
certain scripts from untrusted users.

Signed-off-by: Adam Ludvik <ludvik@bitwise.io>