Load path config and exit on file/directory errors
Adds PathConfig and the load_path_config(), which is called from
validator's main().  Verifies that data_dir and log_dir exist and
are writable.  Also verifies that the identity signing key can be
loaded.

This moves the use of load_identity_signing_key out of Validator's
constructor and into main() so errors are caught earlier.

Adds a command-line option --config-dir to validator.

Signed-off-by: Shawn T. Amundson <amundson@bitwise.io>