Add wait parameter to GET /batch_status endpoint
Add `wait` query parameter to `GET /batch_status` in the REST API.
If set to a number of seconds, the API will wait until all
specified batches are committed to the blockchain, or until the
time has elapsed. If no value is specified, or a non-integer value
other than `false` is specified, the timeout will default to 95%
of the REST API's timeout.

Signed-off-by: Zac Delventhal <delventhalz@gmail.com>