Rearrange config commands into sections
The sawtooth config subcommands are rearranged into meaningful subgroups,
`proposal` and `settings`.

The `proposal` subcommand group, is made up of the `create` subcommand.
This takes the place of the `set` and `propose` commands by combining
them into one operation, which either records a file or submits the
settings to the validator.

The `settings` subcommand group is made up (currently) of the `list`
subcommand. This command displays the list of current on-chain settings,
read from a validator's REST api.

Signed-off-by: Peter Schwarz <peterx.schwarz@intel.com>