Allow for differention of return values by status
Throwing the http error in all case as a CliException makes it hard to
make decisions based on the various response codes that can be returned
from the API.  With leaf nodes, a 404, simply means there is no value at
the address, and we can return None.  With sending batches, we can
return the result on a 202, as that is the response code with this
operation.

Signed-off-by: Peter Schwarz <peterx.schwarz@intel.com>