Fix scheduler finalize/completion bug
The completion status of the scheduler is updated in
set_transaction_execution_result() when the transaction
queue is empty and the scheduler is finalized.  However,
finalize() did not have a comparable update to the
completion status, so if finalize() was last, the
scheduler would never complete.

This fixes that issue and adds a unit test for the
scheduler which verifies the expected behavior.

Signed-off-by: Shawn T. Amundson <amundson@bitwise.io>