Update build_debs for new 0.8 packaging
To add intra-project package dependencies, packages must be built and
installed in a specific order.

There is also a need to set the path for several packages, so rather
than have a bunch of if statements in a for-loop, I removed the for loop
and created a function.

Finally, the build command has been changed so that all packages are
built the same way.

Signed-off-by: Adam Ludvik <ludvik@bitwise.io>