Set the aiohttp version specifically
This applies to pip installs of aiohttp in vagrant and docker.
There is a breaking change in v2+ of the test framework for
aiohttp that breaks our tests. Further work will be to make
our tests compatible with latest aiohttp pip packages and
explicitly build debs based on set versions of pip packages.

Signed-off-by: Boyd Johnson <boydx.johnson@intel.com>