Add --wait flag to sawtooth cli submit command
Add a wait flag to the submit command to specify that the CLI
should wait for the batches to be committed before exiting. The flag
can optionally be set to an integere to specify a timeout in seconds.
If the timeout elapses before all batches have been committed, the
batch statuses will print, and the CLI will exit with a status of 1.
If told to wait with no integer, the CLI will wait indefinitely,
repeatedly querying the REST API if necessary.

Refactor the way POST requests and query parameters are handled by
the CLI's REST Client to satisfy the new use case.

Signed-off-by: Zac Delventhal <delventhalz@gmail.com>