Change test to not name containers explicitly
The shutdown-smoke test container needs to know about the
names of sibling containers so it could send signals and shut
them down. To satisfy this, while also not giving the containers
explicit names, this test was rewritten to start any sibling containers
that it needed and then control those containers throughout the test.

Signed-off-by: Boyd Johnson <boydx.johnson@intel.com>