Fix Dropping batches that have satisfied dependencies
While the publisher is waiting for the last block to be committed,
the publisher was throwing away batches that have dependencies
present in that block. This adds the batches in the block to the
committed_txn_cache so when the dependencies are checked the
soon to be comitted batches are accounted for.

Signed-off-by: Andrea Gunderson <andreax.gunderson@intel.com>