Prevent double processing of recently-seen blocks.
If the chain controller receives a block that is already in its processing block queue, it will not schedule it for verification.  This fixes a bug where a validator receives the same block from multiple validators and if it schedules the second one while the first one is still being verified, an exception occurs on ChainController.on_block_validated() when it attempts to remove the block's entry in the processing blocks queue.

Signed-off-by: Jamie Jason <jamie.jason@intel.com>