Handle decode errors in Block/BatchGetRequests
Changing from pip installs to apt-get installs from the sawtooth repo
makes the tests that intentionally request bad blocks/batches throw a
different error (DecodeError).

I thought this was because of a different version of LMDB, which is why
I added a deb for that to match the version that was being installed by
pip, but that didn't do it. It seems like the DecodeError exception
should be caught anyways, so I added it in.

Signed-off-by: Adam Ludvik <ludvik@bitwise.io>