Cancel the scheduler when finalize never called
The BlockPublisher will start a new scheduler based
on a new state root, after the chain head is updated. The
TransactionExecutor needs to know to stop executing on the
previous scheduler if finalize will never be called. Similarly
the BlockValidator will early return on block validation so cancel
needs to be called on the scheduler at that point.

This fixes the TransactionExecutor threadpool getting exhausted seen
in the poet smoke test when target_wait_time=5.

Signed-off-by: Boyd Johnson <boydx.johnson@intel.com>