Fix getting from block store with wrong id type
Previously when fetching a block from the block store with a valid
batch/txn id, or when fetching a batch/txn with a valid block id,
unpredictable errors would be thrown as the wrong sort of resource
was sent to various processes. These errors were then caught by the
client handlers and sent back to the client as an `INVALID_ID`.

This behavior is inconsistent (will not occur when fetching a batch
with a valid txn id for example), and is better handled within the
block store implementation itself. All invalid ids will now send
back the same errors, and client handlers will report every
failure as a `NO_RESOURCE`.

Signed-off-by: Zac Delventhal <delventhalz@gmail.com>