Add zTest block count to the validator and consensus states.
In order to enforce the zTest policy, we need to know how many blocks in total and for a particular validator have been claimed that are covered by the zTest. This is not the same as the total number of blocks claimed as any blocks that are covered with the fixed duration local mean calculation do not factor into the zTest calculcation.

Signed-off-by: Jamie Jason <jamie.jason@intel.com>