PoET enclave factory uses PoET config view fixed duration block count.
In computing the local mean for a PoET wait timer, the fixed duraction block count is used to determine if the local mean should be simply a ratio of target to initial wait times or a history of wait timer local means.

When creating the PoET enclave and setting the wait timer parameters, get the fixed duration block count from the PoET config view.

Signed-off-by: Jamie Jason <jamie.jason@intel.com>