Update PoET block publisher to short-circuit block initialization failure.
This update changes the block publisher so that in the case of failure to initialize a block (for example, because of PoET policies), it will keep the block ID that was trying to be build upon.  Until the previous block ID changes, the PoET block publisher will immediate return failure as it knows that it cannot build on top of that block.

Signed-off-by: Jamie Jason <jamie.jason@intel.com>