Add Transaction and Batch encodes to JS SDK
Add a module for packaging transactions and batches into
the proper protobufs, either to submit to a validator
(batches), or to a batcher (transactions). A single
client could be their own batcher.

Each encoder has three methods:
  - `create`: Create a new instance of the applicable message
  - `encode`: Encode one or more instances in a serialized list
  - `createEncoded`: Convenience methods that creates an instance,
    then encodes it (will be the only instance in the list).

Also adds a TransactionList protobuf for serializing multiple
Transactions to be transmitted to an external batcher.

Signed-off-by: Zac Delventhal <delventhalz@gmail.com>