Remove non-ballot mode from sawtooth config
Remove the non-ballot or non-voting mode from the sawtooth config
transaction family reference implementation.  This change requires that
there be at least on public key set in the authorized keys list before
any setting can be changed.

The end result is a more secure model for handling configuration changes
on-chain.

Signed-off-by: Peter Schwarz <peterx.schwarz@intel.com>