Add minimal error handling on receiving messages
A common case that has been seen is a differing zmq socket type
has made a connection to the router socket and sent a message, causing
an exception to be thrown and the receive message coroutine to stop
running. Now if the validator receives a message that causes an error
in the Interconnect the error will be logged. In the future we may want
to take action against a connection that is sending the validator malicious
tcp messages.

Signed-off-by: Boyd Johnson <boydx.johnson@intel.com>