Save state deltas on squash
When the context manager is running the squash operation, and it is told
to persist the changes (i.e. all batches/txns have been processed and
their state changes produced), generated and save the series of state
changes for that state root hash.  This should only be done if changes
exist.

Signed-off-by: Peter Schwarz <peterx.schwarz@intel.com>