Fix periodic test failure in smoke tests
In the smoke test, there is a periodic failure, based around the timing
in the system.  The rest api may or may not be ready, the validator may
not have a genesis block, and so forth, before the tests run.  This
commit takes the `_wait_for` from the test_config_smoke, and moves it to a
new package, `integration_tools` as `wait_for_status`.  This function
will give the system time to settle out, if needed, before running the
tests.

A convenience method `wait_for_rest_apis` takes a list of "<host>:port"
values and checks the `/blocks` endpoint for 200 responses.

Signed-off-by: Peter Schwarz <peterx.schwarz@intel.com>