Replace jackson-dataformat-cbor with co.nstant.in.cobr in Intkey Java
This change allows us to encode the intkey state HashMap in
the correct format to work with the python intkey and javscript
intkey examples. Jackson was formatting it to be an indefinite
map instead of a definite map. This caused the state root hash
to be different on validation when running a mixed network.

Signed-off-by: Andrea Gunderson <andreax.gunderson@intel.com>