Split State Delta Subscribe process Into Phases
Split the state delta subscription registration process into phases.
The first phase checks the validitity of the request and sends the
acknowledgement, the later sends adds the subscriber and sends any
"catch-up" events needed by the subscriber.

This change allow subscribers to recieve the appropriate ACK before
receiving events, and simplifies the complexity of clients needing to
buffer events while waiting on that ACK.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>