Replace ProcessPoolExecutor used by sig verify with ThreadPoolExcecutor
As part of a desire to run CPU intensive secp256k1 signature
verification outside of the core validator process, a
ProcessPoolExecutor execution strategy was chosen which uses
multiprocessing to fork and execute the signature verification
methods. Due to the extensive use of threading inside the
validator process, it is quite likely that a thread is inside
a critical section of logging code when the fork occurs. If
the signature verification code makes use of logging, the
logging call will deadlock on the release of the logging lock
inside the child process (because the thread holding the lock
is frozen inside the child process). This was leading to
intermittent failures during testing.

See http://bugs.python.org/issue6721

This change introduces a fix to this issue by replacing the
ProcessPoolExecutor with a ThreadPoolExecutor, ensuring the
calls run within the same process. This is intended to be a
short-term change, as it is still desirable to have the CPU
intensive signature verification work handled by a different
process.

Signed-off-by: James Mitchell <mitchell@bitwise.io>