Implement capability to use a validator config file
This commit adds a ValidatorConfig object, an example
validator config file, and updates the validator cli to
check for a config file and combine it with any command
line options set.

Important to note the public-uri is no longer "required" to
be set on the command line, but if it is not set by either
the command line or the config file the validator will
shutdown.

Signed-off-by: Andrea Gunderson <andreax.gunderson@intel.com>