Update docs and default docker compose file for config changes
In order to be able to make config changes, the genesis block
must now contain the setting for the authorized key that
can submit config changes. Docs and default docker compose
file were updated to account for this, so that the intro doc
will work for both Ubuntu and docker installs.

Signed-off-by: Todd Ojala <todd@bitwise.io>