Update context manager to pass unit tests
The context data structure is updated to have separate methods
for directly setting a value, or setting that value from the
merkle tree. The futures are updated to keep track of if the
value was set in the context. create_context and squash are
updated to do breadth-first search starting with the contexts
that are given to the method. For squash, each context since
the starting state root is accessed. For create_context the
end condition is either all contexts have been accessed, or
the full addresses in the inputs have found the first
context that have that address set. The checking for valid
addresses to set and get from has been relaxed so that if the
address is within a provided namespace, there won't be an
AuthorizationException, but use of namespaces within
inputs and outputs is not fully supported yet.

Signed-off-by: Boyd Johnson <boydx.johnson@intel.com>